// src/views/Lists/Materials/interfaces.ts
export interface ModalTypes {
  HAMID?: string;
  HAMTIPIID?: string;
  HAMTIPIADI?: string;
  KALINLIKID?: string;
  IPLIKCINSIID?: string;
  MAXFIREORANI?: string;
  MINFIREORANI?: string;
  BIRIMID?: string;
  NOTU?: string;
  KALINLIKADI?: string;
  IPLIKCINSIADI?: string;
  VARSAYILANBIRIMADI?: string;
  MAXIMUMFIREORANI?: string;
  MINIMUMFIREORANI?: string;
  }

  export interface IModalForm {
    ADI: string
    HAMID: string
    ADI2: string
    ADI3: string
    ADI4: string
    HAMTIPIID?: string
    HAMTIPIADI?: string
    KALINLIKID?: string
    IPLIKCINSIID?: string
    MAXFIREORANI?: string
    MINFIREORANI?: string
    BIRIMID?: string
    NOTU?: string
    KALINLIKADI?: string
    IPLIKCINSIADI?: string
    VARSAYILANBIRIMADI?: string
    MAXIMUMFIREORANI?: string
    MINIMUMFIREORANI?: string
  } 